feat: style changies, code refactor
