+++ src/Components/Data/Data.css
@@ -5,3 +5,11 @@
   column-gap: 20px;
   row-gap: 20px;
 }
+
+.dataTitle {
+  margin: 0;
+  padding-top: 30px;
+  width: 100%;
+  font-size: 3rem;
+  line-height: 2;
+}
