+++ src/Components/Data/Planet.css
@@ -1,9 +1,21 @@
 .planetWrapper {
-  box-sizing: border-box;
   padding: 10px;
   width: 300px;
+  background-color: black;
+  box-shadow: 0 10px 20px rgba(255, 255, 255, 0.4);
+  box-sizing: border-box;
   border: 2px solid aliceblue;
   border-radius: 10px;
+  transition: transform 0.3s;
+}
+
+.planetWrapper:hover {
+  transform: scale(1.1);
+}
+
+.planetImageContainer {
+  max-width: 100%;
+  height: 100px;
 }
 
 .planetDescription {
