+++ src/Components/Data/PlanetImage.css
@@ -1,0 +1,4 @@
+ .planetImage {
+   max-width: 100%;
+   height: 100px;
+ }
