--- src/Components/PlanetImage.css
@@ -1,4 +1,0 @@
-.planetImage {
-  max-width: 100%;
-  height: 100px;
-}
